requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord-py-interactions==4.3.0
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-decouple==3.6
requests==2.28.1
typing-extensions==4.3.0
urllib3==1.26.10
yarl==1.7.2
